deps: migrate rusqlite 0.32 -> 0.40
rusqlite 0.38 stopped implementing ToSql for usize by default, so the audit purge's LIMIT bind (a usize batch size) no longer compiled — the reason Dependabot #241 failed. Cast it to i64 at the bind site; the public delete_older_than signature is unchanged. Across the 8-version jump this is the only affected call site; column_decltype and trace features are still supported. Bundled SQLite moves 3.45 -> 3.51 (wire/SQL behavior unchanged).

+++ crates/dbflux_storage/src/repositories/audit.rs
@@ -599,7 +599,7 @@ impl AuditRepository {
 
         let deleted = conn.execute(
             "DELETE FROM aud_audit_events WHERE created_at_epoch_ms < ?1 LIMIT ?2",
-            rusqlite::params![cutoff_ms, limit],
+            rusqlite::params![cutoff_ms, limit as i64],
         )?;
 
         Ok(deleted as i64)
